Refactor app to allow for multiple relay plates
The user is now expected to explicitly declare the
(board, relay) address for each zone they wish to enable,
in addition to the number of zones.  This allows for
arbitrary choice of relays for zones and for using as
many boards as the user wishes (up to the physical max).