Add functionality to save a sequence
As far as sequences go, half of CRUD is now available.
Create is up next, then delete.  There is some dirty hard-
coding in here that needs to be cleaned up, and it's worth
mentioning somewhere that hard-coded URLs need to be changed
as well (only just now am I starting to use url_for()).