Implement functionality to run a sequence
There are now routes, globals, and helpers for running a
sequence, plus apposite changes to the sequences template.

The idea is to only have one sequence running at a time,
and the user is disallowed from running zones when there is
an active sequence.

The included sequence with id 0 is intended for testing
sequences on the hardware.  What should remain for this
branch are bugs and stylistic niggles.

Finally, it is worth mentioning to myself that I should
stash any changes needed to run the app in the dev
environment; the current code state should reflect and
continue to reflect a working state (e.g., not importing
pi-plates libraries when running the dev server).