Generalize sequence template into list template
In a further attempt to thin out the numbers of templates,
the sequences list template has been refactored into a list
template which can hopefully be re-used for listing
configuration items and what have you.

Also included in this commit is some basic message
flashing and some cleaned up markup in extant templates
(those not slated for refactoring at least; the edit page
is the next victim).