Generalize edit template
This commit ended up being bigger than I would have liked
owing to requisite alterations to the underlying data
model.  Better to get this kind of thing sorted before it
gets reused.

More changes are probably needed before these templates
can be actually reused but the old functionality is as it
was for the most part before these changes occurred.