Fix newSequence route
Fixing the error introduced by commit 915408b8f6ab8649cbca969097134339b6a29519 whereby the
edit template was generalized such that the form on the page
submits to the same endpoint (i.e., remove the form's action
attribute), and this change was not reflected in the
newSequence route.  This has been rectified.