Implement functionality for scheduling sequences
This is the first stab at including functionality for scheduling
sequences to be run on a weekly basis.

Schedules are the editable objects which the user CRUDs and sets as
either active or inactive.  These objects each hold a job queue, and a
timer thread waiting for the next scheduled sequence to be run.

The jobs only hold data indicating when in the week they should be run
but provide instance methods for determining how much time and when the
next run will be, and jobs can be compared on said. Each job corresponds
to a single run of a single sequence (for now).