Mitigate potential race conditions
If the user sets two schedules to have jobs that run at the same time on
the same day, Python may or may not then schedule those jobs close
enough together to cause multiple sequences to run concurrently;
thereafter the class will exhibit undefined behaviour.  A static lock
has been added to the Sequencer class to mitigate this occurrence.

Still needed is a check in the schedule controller which forbids saving
schedules having jobs scheduled to run at the same time on the same day.
This does not result in undefined behaviour, but the less than useful
behaviour that concurrent jobs in a schedule will not all run; the first
job in the job queue will run, the queue gets rotated, and the next job
(which was supposed to be run just a few microseconds ago) is scheduled
to be run next week.