Refactor keypad template
In order to make the confirmation message a bit more informative, a
rabbit hole was tumbled into whereby the time selection route was
removed in favour of a slider inside a form to indicate the desired time
interval before finalizing the selection by hitting a relay button.
This may or may not remain pending user testing.