Add check for job uniqueness when saving schedules
This (hopefully) comprises mitigation for one or more bugs introduced
when the user is able to save a schedule containing jobs set to run
concurrently.  This would cause erroneous behaviour in the schedule,
likely postponing the next run to at least one week in the future.  Now
the user is instructed to remove concurrently scheduled jobs when
attempting to save a schedule containing said.