Add countdown timer to keypad buttons
Previously, there was no indication as to how much time remains for an
active zone in the keypad view.  This has been partially rectified by
adding a dumb javascript countdown timer and the means for feeding it
the remaining times for zones currently active.  The countdown has a
tendency to lag behind the real time being counted down in python in the
backend, but it's way better than nothing.