Generalize away from sprinklers
This much too large commit represents a movement away from controlling
sprinklers contingent on a pump, and towards controlling relays more
generally.  Relays can now be configured to be dependent on other
relays, which themselves can be dependent on other relays ad infinitum
(until you run out of relays).

Two example configurations are now included, one for a sprinkler system,
and another for a ham radio with disparately located listening stations.
Hopefully the user can come up with other scenarios to leverage this
ability to control and schedule relays with dependencies.

Currently these config files are manually edited to reconfigure the
relays used by the app; still needed is a means for the user to change
the relays connected, their names and dependencies et cetera, in the
web app.