Add unique id's to Relay objects
This probably could have been lumped into the last big commit, but then
it would have been too beautiful to be palatable.  As it should be,
the app assigns and manipulates relays based on a (non-crypto) unique
identifier string, and not names or sequential numbers.

Some json files were moved into a config/ folder to clean up the root;
more of this to come.