Validate form-submitted configuration data
This commit is much too big to be beautiful.  The reasoning behind its
size is that the initial push toward implementing some data validation
immediately led the codebase to a non-working state.  In order to not
push broken functionality that was previously working, a refactoring
effort was necessary, and this commit is the result of said effort.

With form validation in place, the user may now CRUD Boards, Relays,
Sequences, and Schedules.  The edit page still needs a prettier style.

Platelet has been removed, and the components previously marshalled by
it are now singletons marshalling themselves and controlling access to
their data.  These components now inherit from a superclass which
coerces them to state how their data ought to be saved and loaded.

Configurations present in the tree are now blank, so that the user is
not constrained to any preconceptions as to how to configure the system.

The Relay blueprint handles both configuring relays and turning them on
and off manually; this latter functionality ought to be split off to
form a dashboard blueprint.