Constrain schedule editing to inactive schedules
To simplify timer logic for schedules (i.e., in order to not have to
reset timers when a schedule is edited), schedules have been constrained
such that the timer has to be off in order for the user to edit a
schedule.  Also, schedules previously were created in the active state,
which led to some undefined behaviour; creating them in the inactive
state precludes this behaviour.